Validate user-supplied optical system definition
The error handling mechanism in polsim was completely overhauled.
Previously a `Result<T, PolSimError>` would be returned by operations
that could fail, but this has been replaced by the `error_chain` crate.
The motivation for this is to make it easier for me to provide a user
with a series of breadcrumbs leading to an error in the definition of
their optical system.