Normalize elliptically polarized beams
This changes the way that elliptically polarized beams are validated and
reported. Previously, an elliptically polarized beam would be imported
as-is, meaning that it wouldn't be normalized to 1. For consistency,
this is being changed so that all beams start out normalized to 1.
Specifically, the intensity is normalized to 1, and the components are
normalized accordingly.